Refactor unit test implmentation, move components to common
