+++ services/minespace-web/src/components/Forms/projects/projectSummary/Applicant.tsx
@@ -25,7 +25,7 @@ import RenderSelect from "@mds/common/components/forms/RenderSelect";
 import RenderCheckbox from "@mds/common/components/forms/RenderCheckbox";
 import { normalizePhone } from "@common/utils/helpers";
 import { getOrgBookCredential } from "@mds/common/redux/selectors/orgbookSelectors";
-import { PaymentContact } from "./PaymentContact";
+import { PaymentContact } from "@mds/common/components/projectSummary/PaymentContact";
 
 const { Title, Paragraph } = Typography;
 interface IVerifiedCredential {
